final/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
google-cloud-vision==0.39.0
google-cloud-storage==1.19.1
google-cloud-datastore==1.9.0